% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_pcore_data}
\alias{get_pcore_data}
\title{function to download data from the paleocore API}
\usage{
get_pcore_data(dataset = "turkana", version = "v1",
  base_url = "http://paleocore.org", full_related = FALSE, limit = 20,
  ...)
}
\arguments{
\item{dataset}{Name of paleocore dataset you want data from. Default is "turkana"}

\item{version}{Which version of the paleocore to use. Default is "v1"}

\item{base_url}{The base url for the API, with no trailing slash. Defaults to http://paleocore.org}

\item{full_related}{Whether to include full information from related resources or simply return the URI of related resources.  Default is FALSE.}

\item{limit}{Limits the number of records returned.  Default is 20. Use `limit=0` to return all matching records.}

\item{...}{An arbitrary number of filter critera in the following form: genus__contains="Austral"}
}
\description{
This function creates a properly formatted URL string and attempts to get the corresponding data from the Paleocore API
}
\examples{
get_pcore_data(tribe__contains="Tragel", limit=0)
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

