% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getPCdata}
\alias{getPCdata}
\title{function to download data from the paleocore API}
\usage{
getPCdata(project = "turkana", version = "v1", format = "csv", ...)
}
\arguments{
\item{project}{Name of paleocore project you want data from. Default is "turkana"}

\item{version}{Which version of the paleocore to use. Default is "v1"}

\item{format}{The format in which to return data.  Default is "csv" other options include "json" and "xml". If other than CSV, you are responsible for parsing this data into a useful R object (e.g. dataframe).}

\item{limit}{Limits the number of records returned.  Default is 20. Use `limit=0` to return all matching records.}

\item{...}{An arbitrary number of filter critera in the form of "genus__contains=Austral"}
}
\description{
This function creates a properly formatted URL string and attempts to get the corresponding data from the Paleocore API
}
\examples{
getPCdata(tribe__contains="Tragel", limit=0)
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

