% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getPCresources}
\alias{getPCresources}
\title{function to get available PaleoCore resources}
\usage{
getPCresources(version = "v1", base_url = "http://paleocore.org")
}
\arguments{
\item{version}{The version of the API to use, defaults to "v1"}

\item{base_url}{The base url for the API, with no trailing slash. Defaults to http://paleocore.org}
}
\description{
This function gets a listing of the individual data resources that are accessible using the getPCdata() function
}
\examples{
getFilters("turkana")
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

