% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list_pcore_filters}
\alias{list_pcore_filters}
\title{function to get filter field options for a particular API dataset}
\usage{
list_pcore_filters(dataset, version = "v1",
  base_url = "http://paleocore.org")
}
\arguments{
\item{dataset}{The name of the paleocore dataset to get information on}

\item{version}{The version of the API to use, defaults to "v1"}

\item{base_url}{The base url for the API, with no trailing slash. Defaults to http://paleocore.org/}
}
\description{
This function gets a listing of the fields that are available for filtering when using the get_pcore_data() function
}
\examples{
list_pcore_filters("turkana")
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

