% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_pcore_credentials}
\alias{set_pcore_credentials}
\title{simple function to set your paleocore api_key and username as global variables}
\usage{
set_pcore_credentials(username, api_key)
}
\arguments{
\item{username}{Your paleocore username.  Contact an administator to get one}

\item{api_key}{Your api_key.  If you already have a username, you can get an API key at http://paleocore.org/apikey}
}
\description{
This is a convenience function to store your paleocore username and api_key
}
\examples{
set_pcore_credentials("wabarr","asdfk3823kd34983kasdkf013knd83m")
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

