% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setPCcredentials}
\alias{setPCcredentials}
\title{simple function to set your paleocore api_key and username as global variables}
\usage{
setPCcredentials(username, api_key)
}
\arguments{
\item{username}{Your paleocore username.  Contact an administator to get one}

\item{api_key}{Your api_key.  Contact an administator to get one}
}
\description{
This is a convenience function to store your paleocore username and api_key
}
\examples{
setPCcredentials("wabarr","asdfk3823kd34983kasdkf013knd83m")
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

