% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list_pcore_datasets}
\alias{list_pcore_datasets}
\title{function to get available PaleoCore datasets}
\usage{
list_pcore_datasets(version = "v1", base_url = "http://paleocore.org")
}
\arguments{
\item{version}{The version of the API to use, defaults to "v1"}

\item{base_url}{The base url for the API, with no trailing slash. Defaults to http://paleocore.org}
}
\description{
This function gets a listing of the individual data datasets that are accessible using the get_pcore_data() function
}
\examples{
list_pcore_datasets("turkana")
}
\keyword{API}
\keyword{PaleoCore}
\keyword{paleoanthropology}

